--- include/client.h
@@ -1,4 +1,0 @@
-#ifndef CLIENTTHREAD_H
-#define CLIENTTHREAD_H
-#include "../archive/client.c"
-#endif
