--- include/createHeaderBar.h
@@ -1,4 +1,0 @@
-#ifndef CREATEHEADERBAR_H
-#define CREATEHEADERBAR_H
-#include "../src/ui/createHeaderBar.c"
-#endif
