+++ include/getClientSocket.h
@@ -1,0 +1,4 @@
+#ifndef GETCLIENTSOCKET_H
+#define GETCLIENTSOCKET_H
+#include "../src/getClientSocket.c"
+#endif
