+++ include/getServerSocket.h
@@ -1,0 +1,4 @@
+#ifndef GETSERVERSOCKET_H
+#define GETSERVERSOCKET_H
+#include "../src/getServerSocket.c"
+#endif
