--- include/headerMenu.h
@@ -1,4 +1,0 @@
-#ifndef HEADERMENU_H
-#define HEADERMENU_H
-#include "../src/ui/headerMenu.c"
-#endif
