--- include/onDeviceListActivatedOld.h
@@ -1,4 +1,0 @@
-#ifndef ONSIDEBARITEMSELECTED_H
-#define ONSIDEBARITEMSELECTED_H
-#include "../src/ui/onDeviceListActivatedOld.c"
-#endif
