--- include/onRefreshClicked.h
@@ -1,4 +1,0 @@
-#ifndef ONREFRESHCLICKED_H
-#define ONREFRESHCLICKED_H
-#include "../src/ui/onRefreshClicked.c"
-#endif // !ONREFRESHCLICKED_H
