+++ include/receiveFile.h
@@ -1,4 +1,4 @@
 #ifndef RECEIVEFILE_H
 #define RECEIVEFILE_H
-#include "../archive/receiveFile.c"
+#include "../src/receiveFile.c"
 #endif
