+++ include/sendFile.h
@@ -1,4 +1,4 @@
 #ifndef SENDFILE_H
 #define SENDFILE_H
-#include "../archive/sendFile.c"
+#include "../src/sendFile.c"
 #endif
