--- include/server.h
@@ -1,4 +1,0 @@
-#ifndef SERVERTHREAD_H
-#define SERVERTHREAD_H
-#include "../archive/server.c"
-#endif
