--- include/whileDeviceClicked.h
@@ -1,4 +1,0 @@
-#ifndef WHILEDEVICECLICKED_H
-#define WHILEDEVICECLICKED_H
-#include "../src/ui/whileDeviceClicked.c"
-#endif
